2️⃣🧮 update param to double in update_current_phase

+++ packages/client/lib/app/core/modules/collaborator_presence/domain/logic/update_current_phase.dart
@@ -14,7 +14,7 @@ class UpdateCurrentPhase
 
 class UpdateCurrentPhaseParams extends Equatable {
   final bool shouldUpdateCollaboratorsIndex;
-  final int newPhase;
+  final double newPhase;
 
   const UpdateCurrentPhaseParams({
     required this.newPhase,
